changed trade sytem dictionary
before was a dictionary containing a string and a list of trades(dict<string, list<trade>>), and
now is a dictionary containing a string and a dictionary containing a
string and a list of trade (ditc<string, dict<string, list<trade>>>).
This would make it easier to find the buyer and the seller.