added booleans to check trade requests and history
found an issue where the active user woulnd't get any buy requests if
they were not the first key of the "userTrades" dictionary, because the
loop broke when the user were not found in that first lap. Solved with
a boolean to check if the user is part of the dictionary at all, and if
not, then a "no request" messege shows up.
Similar with trade history, if active user was not the first key then a
"no history to show" message showed up. Fixed with a boolean as well.