added "ErrorMsg" method to Functionality class
to show certain messages when an error occurs. The messages have two
strings that would be the "why" of the error and the "what to do next".
Used some keyword for the most common error messages, like "invalid
input" and "press enter to (either continue or go back to previous menu)"