did more fixes and checks to trading system
added some checks so the active user can't accidentally choose
thierselves to trade with, or thier items when selecting another user items to trade.
Added also check to see if an user have items to trade, and if not, is
not possible to select them (it was before the fix).
Added also checks for invalid inputs when selecting index of users or
items.