changed how lines are read from "Trades.csv" file
realized that the items were added to list of items of a trade in th
wrong order, based on which item was first on the general list of items
(userItems). Now corrected that so the items are added in the correct
order. This is important for the next step which is to change the owner
of an item when a trade is acceptethe items are added in the correct
order. This is important for the next step which is to change the owner
of an item when a trade is accepted